Reboot
Finally getting back to this one after a 2-year hiatus. Reorganized some things and reviewed the state that this was left in. Overall architecture is staying, and most of the microcontroller software architecture is staying. However, after taking a few physics classes and more research, the locomotion system will likely be redesigned to use brushed DC motors. We'll see.